requirements.txt
torch
torchvision
kaggle
pandas
numpy
scikit-learn
Pillow
PyYAML
tqdm
pymongo
mongita
fastapi
uvicorn
python-multipart
matplotlib
seaborn
opencv-python
python-dotenv
